refactor(SideBar): Fix entry point
